% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shapes.R
\name{find_shapes}
\alias{find_shapes}
\title{Find Shapefiles}
\usage{
find_shapes(path)
}
\arguments{
\item{path}{Path where to find th files.}
}
\value{
Character vector with shapefile names
}
\description{
Find files within a path which end with ".shp"
}
\examples{
find_shapes("path/to/your/location")
}
