% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes_info.R
\name{get_attributes_info}
\alias{get_attributes_info}
\title{Get attributes summary from Shapefiles}
\usage{
get_attributes_info(path, subfolders = FALSE, report = NULL)
}
\arguments{
\item{path}{\code{character}. Path to either a Shapefile or to a folder.}

\item{subfolders}{\code{logical}. If \code{FALSE} (the default) and \code{path} argument is set to a folder the function won't search within any existing subfolders.}

\item{report}{\code{character}. Path to an Excel to be created which contains all the attributes found. If there are more than one Shapefile the Excel file will have as many tabs as Shapefiles.}
}
\value{
A \code{tibble} object that stores one every Shapefile attributes.
\itemize{
\item \code{shape_name}. Name of the Shapefile.
\item \code{field_name}. Name of each field from a Shapefile.
\item \code{data_type}. Format of each field.
}
}
\description{
Getssearches Shapefiles existing within an existing folder o directory.
}
\examples{
directory <- system.file("extdata", package = "milcao")
get_attributes_info(path = directory)

# Set `subfolder = TRUE` to search within any existing subfolder.
get_attributes_info(path = directory, subfolder = TRUE)
}
