% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_info.R
\name{get_shape_info}
\alias{get_shape_info}
\title{Get vector properties of Shapefile}
\usage{
get_shape_info(path, subfolders = FALSE)
}
\arguments{
\item{path}{\code{character}. Path to either a Shapefile or to a folder.}

\item{subfolders}{\code{logical}. If \code{FALSE} (the default) and \code{path} argument is set to a folder the function won't search within any existing subfolders.}
}
\value{
A \code{tibble} with the following Shapefile properties:
\itemize{
\item \code{shape_name}. Name of the Shapefile.
\item \code{path}. Path to the Shapefile's folder.
\item \code{format}. Name of the file's format.
\item \code{last_edited_gdal}. Shapefile's last edition date.
\item \code{source_encoding}. Character encoding of the Shapefile's attribute table.
\item \code{row_count}. Number of rows of the Shapefile's attribute table.
\item \code{col_count}. Number of cols of the Shapefile's attribute table.
\item \code{geom_type}. Type of geometry of the Shapefile.
\item \code{has_zm}.¿Does the Shapefile has ZM values? Yes (\code{TRUE}) or No (\code{FALSE})
\item \code{crs_type}. Type of coordinates of the CRS (Geographic or Projected).
\item \code{crs_name}. Name of the CRS.
\item \code{crs_epsg}. EPSG code of the CRS.
}
}
\description{
Retrieves vector properties of Shapefiles including file name, geometry type, CRS among other.
}
\examples{
directory <- system.file("extdata", package = "milcao")
get_shape_info(path = directory, subfolder = TRUE)

# Set `subfolder = TRUE` to search within any existing subfolder.
get_shape_info(path = directory, subfolder = TRUE)
}
