% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not-in operator}
\usage{
... \%nin\% NA
}
\arguments{
\item{path}{Path where to find th files.}
}
\description{
Find non-matching elements between two vectors. This is actually a literal negation of the base operator ´\%in\%´
}
\examples{
'comunas' \%nin\% c("Estación Central", "Santiago", "Providencia")
}
