% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shapes.R
\name{find_shapes}
\alias{find_shapes}
\title{Shapefile (.shp) finder}
\usage{
find_shapes(path, full = FALSE)
}
\arguments{
\item{path}{\code{character}. Path to a specific folder.}

\item{full}{\code{logical}. If \code{FALSE} (the default) the function \code{find_shapes()} returns only the base file name of each Shapefile found. If \code{TRUE} it returns the full path name to every Shapefile}
}
\value{
A \code{character} vector containing the names of the Shapefiles found in the folder path.
}
\description{
\code{find_shapes()} searches Shapefiles existing within an existing folder o directory.
}
\examples{
# Error message when using an inexisting path
\dontrun{
find_shapes(path = "github/project/data")

# Error in `find_shapes()`:
# ✖ The path "github/project/data" doesn't exist.
# ℹ Make sure to use an existing folder path.
}
}
