% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes_info.R
\name{get_attributes_info}
\alias{get_attributes_info}
\title{Get attributes summary from Shapefiles}
\usage{
get_attributes_info(path, subfolders = FALSE, report = NULL)
}
\arguments{
\item{path}{\code{character}. Path to either a specific containing Shapefiles or to a Shapefile.}

\item{subfolders}{\code{logical}. If \code{FALSE} (the default) the function \code{get_attributes_info()} won't search within subfolders. If \code{TRUE} it returns all Shapefiles stored within the main folder.}

\item{report}{\code{character}. Path to an Excel to be created which contains all the attributes data found.}
}
\value{
A \code{list} object that stores one dataframe per shapefile attributes.
}
\description{
Getssearches Shapefiles existing within an existing folder o directory.
}
\examples{
# Error message when using an inexisting path
\dontrun{
get_attributes_info(path = "github/project/data/survey.shp")

# Error in `get_attributes_info()`:
# ✖ The path "github/project/data/survey.shp" doesn't exist.
# ℹ Make sure to use an existing Shapefile path.
}
}
